Bucket logic separated out.
