+++ pkg/decision/bucketer/experiment_bucketer.go
@@ -18,40 +18,29 @@
 package bucketer
 
 import (
-	"math"
-
 	"github.com/optimizely/go-sdk/pkg/decision/reasons"
 	"github.com/optimizely/go-sdk/pkg/entities"
-	"github.com/optimizely/go-sdk/pkg/logging"
-	"github.com/twmb/murmur3"
 )
-
-var logger = logging.GetLogger("ExperimentBucketer")
-var maxHashValue = float32(math.Pow(2, 32))
-
-// DefaultHashSeed is the hash seed to use for murmurhash
-const DefaultHashSeed = 1
-const maxTrafficValue = 10000
 
 // ExperimentBucketer is used to bucket the user into a particular entity in the experiment's traffic alloc range
 type ExperimentBucketer interface {
 	Bucket(bucketingID string, experiment entities.Experiment, group entities.Group) (*entities.Variation, reasons.Reason, error)
 }
 
-// MurmurhashBucketer buckets the user using the mmh3 algorightm
-type MurmurhashBucketer struct {
-	hashSeed uint32
+// MurmurhashExperimentBucketer buckets the user using the mmh3 algorightm
+type MurmurhashExperimentBucketer struct {
+	bucketer Bucketer
 }
 
-// NewMurmurhashBucketer returns a new instance of the experiment bucketer
-func NewMurmurhashBucketer(hashSeed uint32) *MurmurhashBucketer {
-	return &MurmurhashBucketer{
-		hashSeed: hashSeed,
+// NewMurmurhashExperimentBucketer returns a new instance of the murmurhash experiment bucketer
+func NewMurmurhashExperimentBucketer(hashSeed uint32) *MurmurhashExperimentBucketer {
+	return &MurmurhashExperimentBucketer{
+		bucketer: MurmurhashBucketer{hashSeed: hashSeed},
 	}
 }
 
 // Bucket buckets the user into the given experiment
-func (b MurmurhashBucketer) Bucket(bucketingID string, experiment entities.Experiment, group entities.Group) (*entities.Variation, reasons.Reason, error) {
+func (b MurmurhashExperimentBucketer) Bucket(bucketingID string, experiment entities.Experiment, group entities.Group) (*entities.Variation, reasons.Reason, error) {
 	if experiment.GroupID != "" && group.Policy == "random" {
 		bucketKey := bucketingID + group.ID
 		bucketedExperimentID := b.bucketToEntity(bucketKey, group.TrafficAllocation)
@@ -75,8 +64,8 @@ func (b MurmurhashBucketer) Bucket(bucketingID string, experiment entities.Exper
 	return nil, reasons.BucketedVariationNotFound, nil
 }
 
-func (b MurmurhashBucketer) bucketToEntity(bucketKey string, trafficAllocations []entities.Range) (entityID string) {
-	bucketValue := b.generateBucketValue(bucketKey)
+func (b MurmurhashExperimentBucketer) bucketToEntity(bucketKey string, trafficAllocations []entities.Range) (entityID string) {
+	bucketValue := b.bucketer.Generate(bucketKey)
 
 	var currentEndOfRange int
 	for _, trafficAllocationRange := range trafficAllocations {
@@ -87,14 +76,4 @@ func (b MurmurhashBucketer) bucketToEntity(bucketKey string, trafficAllocations 
 	}
 
 	return ""
-}
-
-func (b MurmurhashBucketer) generateBucketValue(bucketingKey string) int {
-	hasher := murmur3.SeedNew32(b.hashSeed)
-	if _, err := hasher.Write([]byte(bucketingKey)); err != nil {
-		logger.Error("", err)
-	}
-	hashCode := hasher.Sum32()
-	ratio := float32(hashCode) / maxHashValue
-	return int(ratio * maxTrafficValue)
 }
